+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Nameplate Icons
 
+## [v1.0.1](https://github.com/rbgdevx/nameplate-icons/releases/tag/v1.0.1) (2025-02-23)
+
+- Adding support for Plater, KUI, and ElvUI
+- Removing combat log healer check
+
 ## [v1.0.0](https://github.com/rbgdevx/nameplate-icons/releases/tag/v1.0.0) (2025-01-16)
 
 - Initial release